moving equipmentConnector inside EquipmentType and creating EquipFace
